fix: cast uwrap results to uint256 and drop euint256 support

+++ packages/mock-contracts/contracts/TestBed.sol
@@ -13,14 +13,14 @@ contract TestBed {
 
   function setNumber(InEuint32 memory inNumber) public {
     eNumber = FHE.asEuint32(inNumber);
-    numberHash = euint32.unwrap(eNumber);
+    numberHash = uint256(euint32.unwrap(eNumber));
     FHE.allowThis(eNumber);
     FHE.allowSender(eNumber);
   }
 
   function setNumberTrivial(uint256 inNumber) public {
     eNumber = FHE.asEuint32(inNumber);
-    numberHash = euint32.unwrap(eNumber);
+    numberHash = uint256(euint32.unwrap(eNumber));
     FHE.allowThis(eNumber);
     FHE.allowSender(eNumber);
   }
